requirements.txt
google-genai
pandas
openpyxl
duckduckgo-search
beautifulsoup4
requests
pydantic
python-docx
